+++ tests/Fixtures/DestructorControllerAction/Controller.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+namespace Tests\Fixtures\DestructorControllerAction;
+use Karewan\KnRoute\Attributes\Get;
+class Controller
+{
+	#[Get('/invalid')]
+	public function __destruct() {}
+}